change owner_thread type to int4

+++ src/main/utils_omp.F90
@@ -145,9 +145,9 @@ subroutine limits_omp_work (n1,n2,i1,i2,work,mask,iskip)
 
 end subroutine limits_omp_work
 
-integer(kind=1) function omp_thread_num()
+integer function omp_thread_num()
 #ifdef _OPENMP
- integer(kind=1) :: omp_get_thread_num
+ integer :: omp_get_thread_num
  omp_thread_num = omp_get_thread_num()
 #else
  omp_thread_num = 0
